backend/requirements.txt
fastapi
uvicorn
agno
python-dotenv
openai
sqlalchemy
psycopg2-binary
pgvector
duckduckgo-search
pypdf
python-docx
python-multipart
ddgs
PyJWT
slowapi
prometheus-client
